server/requirements.txt
Flask>=2.1,<3.0
Flask-Cors~=3.0
requests~=2.0
phonenumbers~=8.0

################
# TencentCloud #
################
tls-sig-api-v2>=1.1,<2.0
# TencentCloud SDK: common
tencentcloud-sdk-python-common
# TencentCloud SDK: TRTC only
tencentcloud-sdk-python-trtc
# TencentCloud SDK: SMS only
tencentcloud-sdk-python-sms
